Refactor utils to hooks and remove hardcoded ratio (#259)
