+++ packages/frontend/src/utils/pricer.ts
@@ -583,127 +583,6 @@ export function getMintAndLpPayoffGraph(ethPrice: number) {
   return { leveragePayout, ethPercents, lpPayout }
 }
 
-function getShortParams(ethPrice: number, collatRatio: number) {
-  const squeethIndex = ethPrice ** 2
-  const markRatio = 1.0035
-  const squeethMark = squeethIndex * markRatio
-  const dailyNormFactor = squeethMark / (2 * squeethMark - squeethIndex)
-  const initialCollat = collatRatio * ethPrice
-  const depositValue = initialCollat * ethPrice - squeethMark
-
-  const cuNF0 = dailyNormFactor ** 0
-  const cuNF1 = dailyNormFactor ** 1
-  const cuNF14 = dailyNormFactor ** 14
-  const cuNF28 = dailyNormFactor ** 28
-  const getEthPrices = () => {
-    let inc = Math.floor(ethPrice / 2)
-    return Array(120)
-      .fill(0)
-      .map((_, i) => {
-        if (i === 0) return inc
-        inc += 30
-        return inc
-      })
-  }
-  const getCrabEthPrices = () => {
-    let inc = Math.floor(ethPrice / 2)
-    return Array(100)
-      .fill(0)
-      .map((_, i) => {
-        if (i === 0) return inc
-        inc += 30
-        return inc
-      })
-  }
-
-  return {
-    squeethIndex,
-    markRatio,
-    squeethMark,
-    initialCollat,
-    depositValue,
-    cuNF0,
-    cuNF1,
-    cuNF14,
-    cuNF28,
-    ethPrices: getEthPrices(),
-    crabEthPrices: getCrabEthPrices(),
-  }
-}
-
-export function getSqueethShortPayOffGraph(ethPrice: number, collatRatio: number) {
-  const { markRatio, initialCollat, depositValue, cuNF0, cuNF1, cuNF14, cuNF28, ethPrices, crabEthPrices } =
-    getShortParams(ethPrice, collatRatio)
-
-  const payout0 = ethPrices.map((p) => {
-    return (((-1 * cuNF0 * p ** 2 * markRatio + initialCollat * p) / depositValue - 1) * 100).toFixed(2)
-  })
-
-  const payout1 = crabEthPrices.map((p) => {
-    return (((-1 * cuNF1 * p ** 2 * markRatio + initialCollat * p) / depositValue - 1) * 100).toFixed(2)
-  })
-
-  const payout14 = ethPrices.map((p) => {
-    return (((-1 * cuNF14 * p ** 2 * markRatio + initialCollat * p) / depositValue - 1) * 100).toFixed(2)
-  })
-
-  const payout28 = ethPrices.map((p) => {
-    return (((-1 * cuNF28 * p ** 2 * markRatio + initialCollat * p) / depositValue - 1) * 100).toFixed(2)
-  })
-
-  return {
-    ethPrices,
-    payout0,
-    payout1,
-    payout14,
-    payout28,
-  }
-}
-
-export function getCrabVaultPayoff(ethPrice: number, collatRatio: number) {
-  const { markRatio, initialCollat, depositValue, cuNF0, cuNF14, cuNF28, ethPrices } = getShortParams(
-    ethPrice,
-    collatRatio,
-  )
-
-  const payout0 = ethPrices.map((p) => {
-    return (
-      ((-(cuNF0 + (initialCollat - 2 * cuNF0 * ethPrice) / ethPrice) * p ** 2 * markRatio +
-        (initialCollat + ((initialCollat - 2 * cuNF0 * ethPrice) / ethPrice) * markRatio * ethPrice) * p) /
-        depositValue -
-        1) *
-      100
-    ).toFixed(2)
-  })
-
-  const payout14 = ethPrices.map((p) => {
-    return (
-      ((-(cuNF14 + (initialCollat - 2 * cuNF14 * ethPrice) / ethPrice) * p ** 2 * markRatio +
-        (initialCollat + ((initialCollat - 2 * cuNF14 * ethPrice) / ethPrice) * markRatio * ethPrice) * p) /
-        depositValue -
-        1) *
-      100
-    ).toFixed(2)
-  })
-
-  const payout28 = ethPrices.map((p) => {
-    return (
-      ((-(cuNF28 + (initialCollat - 2 * cuNF28 * ethPrice) / ethPrice) * p ** 2 * markRatio +
-        (initialCollat + ((initialCollat - 2 * cuNF28 * ethPrice) / ethPrice) * markRatio * ethPrice) * p) /
-        depositValue -
-        1) *
-      100
-    ).toFixed(2)
-  })
-
-  return {
-    ethPrices,
-    payout0,
-    payout14,
-    payout28,
-  }
-}
-
 export const getLongChartData = async (days: number, collatRatio: number, volMultiplier: number) => {
   const url = `/api/charts/longchart?days=${days}&collatRatio=${collatRatio}&volMultiplier=${volMultiplier}`
   const response = await fetch(url)
